syncing to head of tfs

+++ Merlin/Main/Languages/Ruby/Libraries.LCA_RESTRICTED/Extensions/ClrBigInteger.cs
@@ -927,7 +927,7 @@ namespace IronRuby.Builtins {
             }
 
             // TODO: Can we do the ToLower in BigInteger?
-            return MutableString.CreateAscii(self.ToString((uint)radix).ToLower());
+            return MutableString.CreateAscii(self.ToString(radix).ToLower());
         }
 
         #endregion
