+++ Merlin/Main/Languages/Ruby/Libraries.LCA_RESTRICTED/Extensions/ClrInteger.cs
@@ -876,7 +876,7 @@ namespace IronRuby.Builtins {
             }
             // TODO: Should we try to use a Fixnum specific ToString?
             // TODO: Can we do the ToLower in BigInteger?
-            return MutableString.CreateAscii(self.ToString((uint)radix).ToLower());
+            return MutableString.CreateAscii(self.ToString(radix).ToLower());
         }
 
         #endregion
